+++ event-logging-base/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module event.logging.base {
+    requires java.xml.bind;
+    requires slf4j.api;
+
+    requires event.logging.api;
+}