+++ event-logging-generator/src/main/java/module-info.java
@@ -1,0 +1,4 @@
+module event.logging.generator {
+    requires java.xml.bind;
+    requires slf4j.api;
+}